requirements.txt
flask
snowflake
requests
pandas
numpy
python-dotenv




